MiMC7 Jupyter Notebook (#141)
@leolara Ready for review.

All MiMC7 notebook files are `part4` of the tutorial, with 4 chapters,
the first 3 of which is fully written out (just because they sort of
came out naturally rather than to an outline first, so I thought this
actually saved more time), and the last of which is an actual outline
with `// TODOs`.

Let me know your thoughts!