Feature/optimise step selector builder (#154)
## Optimisation of Step Selector for Log2(n) Columns

### Description
- **Goal:** is to optimise the Step Selector Builder to have Log2(n)
Columns per n Step Types instead of one Column per each Step Type.
- **Current Implementation:** implemented a new type of Selector and
implemented the build function for it

- **Explanation:** If we use binary representation for the step types
instead of using the actual columns we will get an optimised version for
the Selector Builder since we can get the total of columns from the
following expression
```math
n\_cols = \lceil \log_2(n\_step\_types + 1) \rceil
```
We need take into consideration that we also need to save one binary
value for the case when there is no step type, because of this we do the
math with \(n\_step\_types + 1\).

 ### What is missing
- [x] Ensure compatibility with the backend
- [x] Unit tests

---------

Co-authored-by: nullbitx8 <92404251+nullbitx8@users.noreply.github.com>
Co-authored-by: John Smith <dyxushuai@gmail.com>
Co-authored-by: Leo Lara <leolara@users.noreply.github.com>
Co-authored-by: Jaewon In <soou0710@snu.ac.kr>
Co-authored-by: Steve Wang <qian.wang.wg24@wharton.upenn.edu>