Keep `witness` argument immutable on `halo2_mock_prover` (#175)
Closes #170 

---

On the SuperCircuit's `halo2_mock_prover` method, we are currently
overriding the `super_witness` dict object.
Initially it holds the `TraceWitness` objects of each internal circuit,
but after the loop it holds the JSON representation of the witness of
the circuit.

Since the elements in the dict are indexed by their `rust_id`, after the
loop all the elements have been replaced, leaving the caller of
`halo2_mock_prover` with a "modified" `witness`.

It's not a big problem, but it's confusing sometimes when writing tests,
if you want to verify some of the assignments, because the order of
lines becomes relevant.

I didn't add the return of the `witness_json` because so far I've seen
no use for it. Maybe it could be interesting.