Port Poseidon example to Python (#178)
Closes #101 

---

Based on the Rust implementation, the same circuit is added using the
Python front-end.

Constants are moved to a different file for easier code browsing.

An assertion is added when running the example, to make sure that the
output is the same as the one when using the Rust front-end.