PIL Backend with PIL IR (#165)
Ready for review. @leolara

# High-Level Approach
Convert Chiquito AST to PIL IR and finally to PIL code.

# Future TODOs
- Add support for trace inputs to PIL. Currently, PIL supports user
input via the command line or Rust APIs for the command line. However,
it only supports input for one variable (column). Because Chiquito
circuit might provide inputs for multiple variables via trace (e.g. the
MiMC7 circuit), we can't use use PIL's Rust API as a generic solution
for providing inputs. The only solution is to input statements in the
format of `ISFIRST * ([[input_variable]] - [[input_value]]) = 0`. An
issue with this approach, however, is that our `trace` is too general
and doesn't associate input values to signals, so we might need to
create an alternative `trace` function or mode that does this. All of
the above is due to the fact that PIL only supports automatic witness
inference and cannot feed in external witness.
- Add Python PIL API support for super circuits. Currently we support
single circuit in Python. Should be very do-able, just didn't get the
time to code it up.