Compile to Halo2 middleware (#263)
Switching to a more recent Halo2 version revealed a problem with our
lookups - they don't pass a `lookup-any-sanity-checks` because they are
using only the fixed query. This is why I had to disable the default
features of all halo2 crates (that include `lookup-any-sanity-checks`)
and only enable the rest. I suppose we can refactor our lookups to use
`lookup` instead of `lookup_any`.