Expose debug symbol start and end for the language server (#289)
We have accidentally made the start and end of debug symbol private, so
here's a fix for the language server. Also had to update the halo2 to
the latest version because the language server didn't want to compile
(there was a missing bound on a variable generic inside the halo2
middleware `Expression`, not sure why Chiquito itself compiled
successfully).